# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Webshop API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/product": {
      "post": {
        "tags": [
          "Product"
        ],
        "description": "This is the route for creating a new product",
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/AddProduct"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "description": "This is the route for updating a product",
        "parameters": [
          {
            "name": "updatedProduct",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/EditProduct"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/product/{product_id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "description": "This is the route for getting a specific product",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "description": "This is the route for deleting a product (if possible, else if in use archives it)",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": [
          "Product"
        ],
        "description": "This is the route for getting all products",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AddProduct": {
      "type": "object",
      "properties": {
        "product": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "example": "cool cape"
            },
            "price": {
              "type": "number",
              "example": 1800
            },
            "size": {
              "type": "string",
              "example": "l"
            },
            "description": {
              "type": "string",
              "example": "totally awesome cape"
            },
            "is_archived": {
              "type": "boolean",
              "example": false
            },
            "brands_brand_id": {
              "type": "number",
              "example": 2
            },
            "colors_color_id": {
              "type": "number",
              "example": 2
            },
            "manufacturers_manufacturer_id": {
              "type": "number",
              "example": 2
            },
            "categories_category_id": {
              "type": "number",
              "example": 2
            }
          }
        }
      }
    },
    "EditProduct": {
      "type": "object",
      "properties": {
        "product": {
          "type": "object",
          "properties": {
            "product_id": {
              "type": "number",
              "example": 26
            },
            "name": {
              "type": "string",
              "example": "extra cool cape"
            },
            "price": {
              "type": "number",
              "example": 1111
            },
            "size": {
              "type": "string",
              "example": "l"
            },
            "description": {
              "type": "string",
              "example": "todtally mega awesome cape"
            },
            "is_archived": {
              "type": "boolean",
              "example": false
            },
            "brands_brand_id": {
              "type": "number",
              "example": 2
            },
            "colors_color_id": {
              "type": "number",
              "example": 2
            },
            "manufacturers_manufacturer_id": {
              "type": "number",
              "example": 2
            },
            "categories_category_id": {
              "type": "number",
              "example": 2
            }
          }
        }
      }
    }
  }
}